"Move files into app folder and add route for confirmation page + install react-share"
