+++ package.json
@@ -22,6 +22,7 @@
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-helmet": "^6.1.0",
+    "react-share": "^4.2.1",
     "react-timer-hook": "^2.0.2",
     "smooth-scroll": "^16.1.3"
   },
