+++ src/components/profile/confirmation.js
@@ -1,23 +1,22 @@
 import React from "react"
 
-
 import Layout from "../../components/layout/layout"
-import Header from "../../components/header/header"
+
 import SEO from "../../components/seo"
 import BgImg from "../../images/eoi-bg.svg"
 import High5 from "../../images/mini-thanku.svg"
 
 import "../../styling/survey/thankyou.css"
 
-const AdminThankYou = props => {
+const Confirmation = props => {
   return (
     <Layout>
-      <Header />
+
       <SEO title="Thank you page" />
 
       <h1 className="thanku-title">Thank you!</h1>
       <section className="thanku-container">
-<img src={BgImg} alt="background design" className="bg-img9" />
+      <img src={BgImg} alt="background design" className="bg-img9" />
         <img src={High5} alt="rocket icon" className="high-five" />
         <p className="thanku-message">
           {" "}
@@ -29,4 +28,4 @@ const AdminThankYou = props => {
     </Layout>
   )
 }
-export default AdminThankYou
+export default Confirmation
