+++ src/pages/index.js
@@ -2,7 +2,7 @@
 import React from "react"
 import Layout from "../components/layout/layout"
 import SEO from "../components/seo"
-import "../styling/organisation/index.css"
+import "../styling/app/index.css"
 
 import MiniNavbar from "../components/miniSite/miniNavbar"
