+++ src/styling/app/mininavbar.css
